requirements.txt
pandas
openpyxl
fpdf
pyyaml
schedule
python-dotenv
matplotlib
seaborn
xlsxwriter